+++ tests/test_actions/test_actions.py
@@ -1,6 +1,9 @@
+from collections import OrderedDict
+from unittest import mock
+
 from django.test import TestCase
 
-from dbentry.actions.actions import add_cls_attrs
+from dbentry.actions.actions import add_cls_attrs, summarize
 
 
 def func0():
@@ -9,8 +12,6 @@ def func0():
 
 def func1():
     pass
-
-
 func1.short_description = 'Foo'  # noqa
 func1.allowed_permissions = 'Bar'
 
@@ -46,3 +47,19 @@ class TestAddClsAttrsDecorator(TestCase):
         decorated_func = add_cls_attrs(self.view_class)(func1)
         self.assertEqual(decorated_func.short_description, 'Foo')
         self.assertEqual(decorated_func.allowed_permissions, 'Bar')
+
+
+class TestSummarize(TestCase):
+
+    def test_summarize(self):
+        with mock.patch('dbentry.actions.actions.get_summaries') as get_summaries_mock:
+            get_summaries_mock.return_value = [
+                OrderedDict(ID='1234', Name='Egg & Bacon'),
+                OrderedDict(ID='5678', Name='Spam & Sausage'),
+            ]
+            response = summarize(model_admin=None, request=None, queryset=None)
+        expected = (
+            b'<p>ID: 1234</p><p>Name: Egg & Bacon</p><hr style="break-after:page;">'
+            b'<p>ID: 5678</p><p>Name: Spam & Sausage</p><hr style="break-after:page;">'
+        )
+        self.assertEqual(response.content, expected)